AllProsumerInfo izmena
Vraca min i max potrosnju, proizvodnju i broj uredjaja u datom trenutku. Ispravljeno kod fejkovanja podataka da se vrednosti ne randomizuju za solarne panele ukoliko je noc.

+++ src/backApp/API/API/Repositories/DeviceRepository/IDeviceRepository.cs
@@ -19,7 +19,7 @@ namespace API.Repositories.DeviceRepository
         public Task<EnumCategory.DeviceCatergory> getDeviceCategoryEnum(string idDevice);
         public Task<Prosumer> GetProsumer(string id);
         public Task<List<Prosumer>> GetProsumers();
-        public Task<double> ProsumerDeviceCount(string id);
+        public Task<int> ProsumerDeviceCount(string id);
         public Task EditDevice(string IdDevice, string model, string DeviceName, string IpAddress, bool dsoView, bool dsoControl);
         public Task<Boolean> DeleteDevice(string idDevice);
         public Task InsertLink(ProsumerLink link);
